Support decimal font sizes (such as 8.6 pt) in KFontDialog
QFont already had support for fractional font sizes since Qt 2.
Applications that do not use QFont::pointSizeF() but pointSize()
will get the rounded size. Most applications, however, simply
request a font, use it, or store it using KConfig, so they will
support fractional font sizes transparently.

http://reviewboard.kde.org/r/904/

BUG: 122618

svn path=/trunk/KDE/kdelibs/; revision=1002561

+++ kdeui/fonts/kfontchooser.h
@@ -257,7 +257,7 @@ private:
     Q_PRIVATE_SLOT(d, void _k_style_chosen_slot(const QString&))
     Q_PRIVATE_SLOT(d, void _k_displaySample(const QFont &font))
     Q_PRIVATE_SLOT(d, void _k_showXLFDArea(bool))
-    Q_PRIVATE_SLOT(d, void _k_size_value_slot(int))
+    Q_PRIVATE_SLOT(d, void _k_size_value_slot(double))
 };
 
 Q_DECLARE_OPERATORS_FOR_FLAGS( KFontChooser::DisplayFlags )
